rearranged some things
