+++ Final Project.cs
@@ -1,22 +1,6 @@
-﻿public abstract class Cards
+﻿private static void Main()
 {
-    string name;
-    public Cards(string Name)
-    {
-        Name = name;
-    }
-    public string Name
-    {
-        get => name;
-        set
-        {
-            if (string.IsNullOrEmpty(value))
-            {
-                throw new ArgumentNullException(nameof(value));
-            }
-            name = value;
-        }
-    }
+    Console.WriteLine("Hello World");
 }
